Rename user file to match hooks
