+++ src/views/Root.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { BrowserRouter as Router } from 'react-router-dom'
-import { UserProvider } from 'src/hooks/user'
+import { UserProvider } from 'src/hooks/useUser'
 import App from './App'
 
 const Root: React.SFC = () => (
